WIP: document proposed fix
This is only a partial fix - we still need to run runtime.KeepAlive()
everywhere (but I didn't want to commit that since its a large change
and this is really the crux of it).

TLDR: the C objects this lib wraps share *way* more memory than the
author realized. Its totally busted.